+++ Changelog.md
@@ -31,6 +31,7 @@
 - Fix: The Username Logger sometimes cannot be activated by a short redstone signal
 - Fix: The Codebreaker sometimes loses durability when rightclicking a block even though no codebreaking attempt has been made
 - Fix: List module messages of the Display Case are not translated properly
+- Fix: Taking an item out of a Display Case does not work when the offhand is not empty
 
 --------------------------Changelog for v1.9.4 of SecurityCraft--------------------------